For the mouse click event, decrease the threshold from 1s to 100ms for avoiding registering the same click twice.
Testing has proved that if there is a corner (aka polygon) point at the same location (or even close, seemingly) as a POI, then when clicking on it, the gMapControl1_onMarkerClick event will get called twice. This is bad for generating reports because it doesn't give you the opportunity to select the other corner point.