Adds checkbox to show/hide "suspect readings".
This was implemented a little different than the plain old show/hide for markers, routes, and polygons. Those things use an already existing property in GmapControl / GMapOverlay. I just made the GUI items that toggled them. For this, in POI::UPdateOverlay(), we're just choosing which items to add to the overlay. A side effect is that I had to update FlightPlanner, also. Keep in mind that if you have new contacts indicate as being "suspect" and you're hiding "suspect readings", you won't see the new, suspect ones show up.